Split functions to their own files
