+++ packages/@divvi/mobile/src/public/types.tsx
@@ -223,3 +223,4 @@ export type NetworkId =
   | 'polygon-pos-amoy'
   | 'base-mainnet'
   | 'base-sepolia'
+
